try unify cpu and llvm
